2015/5/12

+++ p2p-parent/p2p-service/src/main/java/com/herongwang/p2p/service/impl/repayplan/RepayPlanServiceImpl.java
@@ -90,7 +90,7 @@ public class RepayPlanServiceImpl implements IRepayPlanService
             //统计所有还款总价格
             for (RepayPlanEntity repayPlanEntity : planlist)
             {
-                monthAmount.add(repayPlanEntity.getMonthAmount()).doubleValue();
+                monthAmount.add(repayPlanEntity.getMonthAmount());
             }
         }
         catch (ServiceException e)
